# front/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "新闻资讯发布与推荐系统 API文档",
    "description": "我的新闻资讯系统后台接口文档",
    "contact": {
      "name": "Tong",
      "email": "tong@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/system/sync": {
      "put": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "edit_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedisSyncTemp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "add_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedisSyncTemp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/role": {
      "put": {
        "tags": [
          "role-controller"
        ],
        "operationId": "edit_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "role-controller"
        ],
        "operationId": "add_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/sync/export": {
      "post": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "export_3",
        "parameters": [
          {
            "name": "redisSyncTemp",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RedisSyncTemp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/system/role/export": {
      "post": {
        "tags": [
          "role-controller"
        ],
        "operationId": "export_5",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/system/user/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getInfo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/tagRel/{id}": {
      "get": {
        "tags": [
          "news-tag-rel-controller"
        ],
        "operationId": "getInfo_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/tag/{id}": {
      "get": {
        "tags": [
          "news-tag-controller"
        ],
        "operationId": "getInfo_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/sync/{newsId}": {
      "get": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "getInfo_3",
        "parameters": [
          {
            "name": "newsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/sync/list": {
      "get": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "list_3",
        "parameters": [
          {
            "name": "redisSyncTemp",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RedisSyncTemp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataInfo"
                }
              }
            }
          }
        }
      }
    },
    "/system/statistics/{id}": {
      "get": {
        "tags": [
          "news-statistics-controller"
        ],
        "operationId": "getInfo_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/role/{id}": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "getInfo_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/role/list": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "list_5",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataInfo"
                }
              }
            }
          }
        }
      }
    },
    "/system/operlog/{id}": {
      "get": {
        "tags": [
          "sys-operation-log-controller"
        ],
        "operationId": "getInfo_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/{id}": {
      "get": {
        "tags": [
          "新闻资讯"
        ],
        "operationId": "getInfo_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/like/{id}": {
      "get": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "getInfo_8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/config/{id}": {
      "get": {
        "tags": [
          "sys-config-controller"
        ],
        "operationId": "getInfo_9",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/comment/{id}": {
      "get": {
        "tags": [
          "评论管理"
        ],
        "operationId": "getInfo_10",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/category/{id}": {
      "get": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "getInfo_11",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/browse/{id}": {
      "get": {
        "tags": [
          "user-browse-controller"
        ],
        "operationId": "getInfo_12",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/sync/{newsIds}": {
      "delete": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "remove_3",
        "parameters": [
          {
            "name": "newsIds",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/role/{ids}": {
      "delete": {
        "tags": [
          "role-controller"
        ],
        "operationId": "remove_5",
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AjaxResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "success": {
            "type": "boolean"
          },
          "warn": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        },
        "additionalProperties": {
          "type": "object"
        }
      },
...
      "RedisSyncTemp": {
        "type": "object",
        "properties": {
          "createBy": {
            "type": "string",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateBy": {
            "type": "string",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "remark": {
            "type": "string",
            "refType": null
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "refType": null
          },
          "newsId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "type": {
            "type": "string",
            "refType": null
          },
          "value": {
            "type": "string",
            "refType": null
          }
        }
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "createBy": {
            "type": "string",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateBy": {
            "type": "string",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "remark": {
            "type": "string",
            "refType": null
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "refType": null
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "roleName": {
            "type": "string",
            "refType": null
          },
          "roleCode": {
            "type": "string",
            "refType": null
          },
          "isDeleted": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
...
      "TableDataInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          }
        }
      }
    }
  }
}